requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-frontmatter==1.0.0
markdown==3.6
nltk==3.8.1
